chore(site): rename and merge utilities
