+++ apps/site/client-context.ts
@@ -1,7 +1,7 @@
 import { cache } from 'react';
 
 import type { ClientSharedServerContext } from '#site/types';
-import { assignClientContext } from '#site/util/assignClientContext';
+import { assignClientContext } from '#site/util/context';
 
 // This allows us to have Server-Side Context's of the shared "contextual" data
 // which includes the frontmatter, the current pathname from the dynamic segments
