+++ apps/site/components/Blog/BlogPostCard/index.tsx
@@ -6,7 +6,7 @@ import FormattedTime from '#site/components/Common/FormattedTime';
 import Link from '#site/components/Link';
 import WithAvatarGroup from '#site/components/withAvatarGroup';
 import type { BlogCategory } from '#site/types';
-import { mapBlogCategoryToPreviewType } from '#site/util/blogUtils';
+import { mapBlogCategoryToPreviewType } from '#site/util/blog';
 
 import styles from './index.module.css';
