+++ apps/site/components/Downloads/DownloadLink.tsx
@@ -5,9 +5,8 @@ import type { FC, PropsWithChildren } from 'react';
 import LinkWithArrow from '#site/components/LinkWithArrow';
 import { useClientContext } from '#site/hooks';
 import type { NodeRelease } from '#site/types';
-import type { DownloadKind } from '#site/util/getNodeDownloadUrl';
-import { getNodeDownloadUrl } from '#site/util/getNodeDownloadUrl';
-import { getUserPlatform } from '#site/util/getUserPlatform';
+import { getNodeDownloadUrl, type DownloadKind } from '#site/util/download';
+import { getUserPlatform } from '#site/util/ua';
 
 type DownloadLinkProps = { release: NodeRelease; kind?: DownloadKind };
