+++ apps/site/components/Downloads/Release/OperatingSystemDropdown.tsx
@@ -8,11 +8,7 @@ import type { FC } from 'react';
 import { useClientContext } from '#site/hooks';
 import { ReleaseContext } from '#site/providers/releaseProvider';
 import type { UserOS } from '#site/types/userOS';
-import {
-  nextItem,
-  OPERATING_SYSTEMS,
-  parseCompat,
-} from '#site/util/downloadUtils';
+import { nextItem, OPERATING_SYSTEMS, parseCompat } from '#site/util/download';
 
 type OperatingSystemDropdownProps = { exclude?: Array<UserOS> };
