+++ apps/site/components/Downloads/Release/PackageManagerDropdown.tsx
@@ -7,11 +7,7 @@ import type { FC } from 'react';
 
 import { ReleaseContext } from '#site/providers/releaseProvider';
 import type { PackageManager } from '#site/types/release';
-import {
-  nextItem,
-  PACKAGE_MANAGERS,
-  parseCompat,
-} from '#site/util/downloadUtils';
+import { nextItem, PACKAGE_MANAGERS, parseCompat } from '#site/util/download';
 
 const PackageManagerDropdown: FC = () => {
   const release = useContext(ReleaseContext);
