+++ apps/site/components/Downloads/Release/PlatformDropdown.tsx
@@ -8,8 +8,8 @@ import { useEffect, useContext, useMemo } from 'react';
 import { useClientContext } from '#site/hooks';
 import { ReleaseContext } from '#site/providers/releaseProvider';
 import type { UserPlatform } from '#site/types/userOS';
-import { PLATFORMS, nextItem, parseCompat } from '#site/util/downloadUtils';
-import { getUserPlatform } from '#site/util/getUserPlatform';
+import { PLATFORMS, nextItem, parseCompat } from '#site/util/download';
+import { getUserPlatform } from '#site/util/ua';
 
 const PlatformDropdown: FC = () => {
   const { architecture, bitness } = useClientContext();
