+++ apps/site/components/withAvatarGroup.tsx
@@ -5,7 +5,7 @@ import type { ComponentProps, FC } from 'react';
 
 import Link from '#site/components/Link';
 import type { AuthorProps } from '#site/types';
-import { getAuthors } from '#site/util/authorUtils';
+import { getAuthors } from '#site/util/author';
 
 type WithAvatarGroupProps = Omit<
   ComponentProps<typeof AvatarGroup>,
