+++ apps/site/components/withBadgeGroup.tsx
@@ -3,7 +3,7 @@ import type { FC } from 'react';
 
 import Link from '#site/components/Link';
 import { siteConfig } from '#site/next.json.mjs';
-import { dateIsBetween } from '#site/util/dateUtils';
+import { dateIsBetween } from '#site/util/date';
 
 const WithBadgeGroup: FC<{ section: string }> = ({ section }) => {
   const badge = siteConfig.websiteBadges[section];
