+++ apps/site/util/api.ts
@@ -1,17 +1,17 @@
-import semVer from 'semver';
+import { satisfies } from 'semver';
 
 import { DOCS_URL, DIST_URL } from '#site/next.constants.mjs';
 
 export const getNodeApiLink = (version: string) => {
-  if (semVer.satisfies(version, '>=0.3.1 <0.5.1')) {
+  if (satisfies(version, '>=0.3.1 <0.5.1')) {
     return `${DOCS_URL}${version}/api/`;
   }
 
-  if (semVer.satisfies(version, '>=0.1.14 <0.3.1')) {
+  if (satisfies(version, '>=0.1.14 <0.3.1')) {
     return `${DOCS_URL}${version}/api.html`;
   }
 
-  return semVer.satisfies(version, '>=1.0.0 <4.0.0')
+  return satisfies(version, '>=1.0.0 <4.0.0')
     ? `https://iojs.org/dist/${version}/docs/api/`
     : `${DIST_URL}${version}/docs/api/`;
 };
